Relocate classes into root folder
