+++ samples/java_springboot/23.facebook-events/src/main/java/com/microsoft/bot/sample/facebookevents/FacebookOptin.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MT License.
 
-package com.microsoft.bot.sample.facebookevents.facebookmodel;
+package com.microsoft.bot.sample.facebookevents;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
