+++ samples/java_springboot/23.facebook-events/src/main/java/com/microsoft/bot/sample/facebookevents/FacebookSender.java
@@ -1,22 +1,22 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MT License.
 
-package com.microsoft.bot.sample.facebookevents.facebookmodel;
+package com.microsoft.bot.sample.facebookevents;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
- * Defines a Facebook recipient.
+ * Defines a Facebook sender.
  */
-public class FacebookRecipient {
+public class FacebookSender {
 
     @JsonProperty(value = "id")
     @JsonInclude(JsonInclude.Include.NON_EMPTY)
     private String id;
 
     /**
-     * The Facebook Id of the recipient.
+     * The Facebook Id of the sender.
      * @return the Id value as a String.
      */
     public String getId() {
@@ -24,7 +24,7 @@ public class FacebookRecipient {
     }
 
     /**
-     * The Facebook Id of the recipient.
+     * The Facebook Id of the sender.
      * @param withId The Id value.
      */
     public void setId(String withId) {
